refactor: rename method

+++ lib/hledger_forecast/cli.rb
@@ -1,4 +1,6 @@
 module HledgerForecast
+  # The Command Line Interface for the application
+  # Takes user arguments and translates them into actions
   class Cli
     def self.run(args)
       end_date = args[:end_date]
@@ -6,27 +8,33 @@ module HledgerForecast
       forecast = File.read(args[:forecast_file])
       transactions = args[:transactions_file] ? File.read(args[:transactions_file]) : nil
 
+      # Output the summary
       return HledgerForecast::Summarize.generate(forecast) if args[:summarize]
 
-      puts "[Using default dates: #{start_date} to #{end_date}]" if args[:default_dates]
+      # Generate the forecast
+      unless args[:skip]
 
-      transactions = Generator.create_journal_entries(transactions, forecast, start_date, end_date)
+        puts "[Using default dates: #{start_date} to #{end_date}]" if args[:default_dates]
 
-      output_file = args[:output_file]
-      if File.exist?(output_file) && !args[:force]
-        print "File '#{output_file}' already exists. Overwrite? (y/n): "
-        overwrite = gets.chomp.downcase
+        transactions = Generator.generate(transactions, forecast, start_date, end_date)
 
-        if overwrite == 'y'
-          File.write(output_file, transactions)
-          puts "File '#{output_file}' has been overwritten."
+        output_file = args[:output_file]
+        if File.exist?(output_file) && !args[:force]
+          print "File '#{output_file}' already exists. Overwrite? (y/n): "
+          overwrite = gets.chomp.downcase
+
+          if overwrite == 'y'
+            File.write(output_file, transactions)
+            puts "File '#{output_file}' has been overwritten."
+          else
+            puts "Operation aborted. File '#{output_file}' was not overwritten."
+          end
         else
-          puts "Operation aborted. File '#{output_file}' was not overwritten."
+          File.write(output_file, transactions)
+          puts "File '#{output_file}' has been created."
         end
-      else
-        File.write(output_file, transactions)
-        puts "File '#{output_file}' has been created."
       end
+
     end
   end
 end
